# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: bladedr API
  version: 1.0.0
  description: Control-plane API for agentless scans, eBPF sensor events, triage, export, and administration.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/v1/login:
    post:
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200': {description: Authenticated, content: {application/json: {schema: {$ref: '#/components/schemas/LoginResponse'}}}}
        '401': {$ref: '#/components/responses/Error'}
...
  /api/v1/me/password:
    post:
      operationId: changeOwnPassword
      description: >-
        Change the caller's own password. Available to every role. While an account has
        must_change_password set, this is the only route it may use; a 403 elsewhere
        means this call has to happen first.
      requestBody: {$ref: '#/components/requestBodies/JSON'}
      responses:
        '204': {description: Password changed}
        '400': {$ref: '#/components/responses/Error'}
        '401': {description: Current password incorrect}
  /api/v1/me/mfa/setup:
    post: {operationId: setupMFA, requestBody: {$ref: '#/components/requestBodies/JSON'}, responses: {'200': {description: Pending TOTP enrollment}}}
  /api/v1/me/mfa/confirm:
    post: {operationId: confirmMFA, requestBody: {$ref: '#/components/requestBodies/JSON'}, responses: {'204': {description: MFA enabled}}}
  /api/v1/me/mfa:
    delete: {operationId: disableMFA, requestBody: {$ref: '#/components/requestBodies/JSON'}, responses: {'204': {description: MFA disabled}}}
  /api/v1/hosts:
    get:
      operationId: listHosts
      parameters:
        - {name: tag, in: query, schema: {type: string}, description: Repeatable key=value filter}
      responses: {'200': {description: Hosts, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Host'}}}}}}
    post:
      operationId: createHost
      requestBody: {$ref: '#/components/requestBodies/JSON'}
      responses: {'201': {description: Host created, content: {application/json: {schema: {$ref: '#/components/schemas/Host'}}}}}
  /api/v1/hosts/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get: {operationId: getHost, responses: {'200': {description: Host, content: {application/json: {schema: {$ref: '#/components/schemas/Host'}}}}, '404': {$ref: '#/components/responses/Error'}}}
    patch: {operationId: updateHost, requestBody: {$ref: '#/components/requestBodies/JSON'}, responses: {'200': {description: Host updated}}}
    delete: {operationId: deleteHost, responses: {'204': {description: Host deleted}}}
...
  /api/v1/hosts/{id}/sensor-tokens:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get: {operationId: listSensorTokens, responses: {'200': {description: Sensor token metadata}}}
    post: {operationId: createSensorToken, requestBody: {$ref: '#/components/requestBodies/JSON'}, responses: {'201': {description: Token returned once}}}
  /api/v1/hosts/{id}/sensor-tokens/{token}:
    parameters: [{$ref: '#/components/parameters/Id'}, {name: token, in: path, required: true, schema: {type: string, format: uuid}}]
    delete: {operationId: revokeSensorToken, responses: {'204': {description: Token revoked}}}
...
  /api/v1/users:
    get: {operationId: listUsers, responses: {'200': {description: Users}}}
    post: {operationId: createUser, requestBody: {$ref: '#/components/requestBodies/JSON'}, responses: {'201': {description: User created}}}
  /api/v1/users/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    patch: {operationId: updateUser, requestBody: {$ref: '#/components/requestBodies/JSON'}, responses: {'200': {description: User updated}}}
    delete: {operationId: deleteUser, responses: {'204': {description: User deleted}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: opaque}
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string}}
...
  requestBodies:
    JSON:
      required: true
      content: {application/json: {schema: {type: object, additionalProperties: true}}}
  responses:
    Error:
      description: Error
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
  schemas:
    Error:
      type: object
      required: [error]
      properties: {error: {type: string}}
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: {type: string}
        password: {type: string, format: password}
        otp: {type: string, pattern: '^[0-9]{6}$'}
    LoginResponse:
      type: object
      required: [token, username, role]
      properties:
        token: {type: string}
        username: {type: string}
        role: {type: string, enum: [admin, operator, viewer]}
...
    Host:
      type: object
      required: [id, mode, status]
      properties:
        id: {type: string, format: uuid}
        hostname: {type: string}
        primary_ip: {type: string}
        ssh_port: {type: integer}
        credential_id: {type: string, format: uuid}
        ssh_host_key: {type: string}
        os_name: {type: string}
        os_version: {type: string}
        kernel: {type: string}
        arch: {type: string, enum: [amd64, arm64]}
        mode: {type: string, enum: [scan_only, scan_plus_sensor]}
        status: {type: string, enum: [pending, online, unreachable, disabled]}
        tags: {type: object, additionalProperties: {type: string}}
        last_seen: {type: [string, 'null'], format: date-time}
        created_at: {type: string, format: date-time}
...